# fix(sumologicexporter): avoid allocations in compressor