--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore: upgrade OT core to 0.76.1 [#1112]
 
 [#1112]: https://github.com/SumoLogic/sumologic-otel-collector/pull/1112
-[0.76.1-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.75.0-sumo-0...0.76.1-sumo-0
+[0.76.1-sumo-0]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.75.0-sumo-0...v0.76.1-sumo-0
 
 ## [v0.75.0-sumo-0]