Avoid an additional copy when compressing data. In most cases, if the io.Reader used implements additional interfaces, a temporary buffer between the reader and the compressing Writer isn't necessary.

I've also fixed the compression benchmark to only measure compression, and not decompression. I've increased the size of the test data to be more representative of real-world workloads as well.

Here's the difference between main and this branch, running the new benchmark:
```
goos: linux
goarch: amd64
pkg: github.com/SumoLogic/sumologic-otel-collector/pkg/exporter/sumologicexporter
cpu: AMD Ryzen 7 2700X Eight-Core Processor         
                       │ benchmark.main.txt │         benchmark.HEAD.txt          │
                       │       sec/op       │   sec/op     vs base                │
Compression/deflate-16          613.5µ ± 2%   382.8µ ± 3%  -37.61% (p=0.000 n=10)
Compression/gzip-16             689.5µ ± 2%   363.7µ ± 3%  -47.25% (p=0.000 n=10)
geomean                         650.4µ        373.1µ       -42.63%

                       │ benchmark.main.txt │          benchmark.HEAD.txt          │
                       │        B/op        │     B/op      vs base                │
Compression/deflate-16        1535.3Ki ± 0%   512.2Ki ± 0%  -66.64% (p=0.000 n=10)
Compression/gzip-16           1535.7Ki ± 0%   512.4Ki ± 0%  -66.63% (p=0.000 n=10)
geomean                        1.499Mi        512.3Ki       -66.64%

                       │ benchmark.main.txt │         benchmark.HEAD.txt         │
                       │     allocs/op      │ allocs/op   vs base                │
Compression/deflate-16          23.000 ± 0%   4.000 ± 0%  -82.61% (p=0.000 n=10)
Compression/gzip-16             23.000 ± 0%   4.000 ± 0%  -82.61% (p=0.000 n=10)
geomean                          23.00        4.000       -82.61%
```